actually, it was close to running. just needed to roll back the overly ambitious changes to graphlayout. up and running. may be some bugs on the horizon
