--- GraphSynthSourceFiles/GraphSynth.GraphLayoutBasic/GraphSynth.GraphLayoutBasic.csproj
@@ -1,12 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.1</TargetFramework>
-    <RootNamespace>GraphSynth</RootNamespace>
-    <Configurations>WPF GUI Debug;WPF GUI Release</Configurations>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\GraphSynth.Base\GraphSynth.Base.csproj" />
-  </ItemGroup>
-</Project>
