+++ GraphSynthSourceFiles/GraphSynth/GraphSynth.csproj
@@ -14,6 +14,7 @@
   <ItemGroup>
     <ProjectReference Include="..\GraphSynth.Base\GraphSynth.Base.csproj" />
     <ProjectReference Include="..\GraphSynth.CustomControls\GraphSynth.CustomControls.csproj" />
+    <ProjectReference Include="..\WpfCustomControlLibrary1\GraphSynth.GraphLayoutBasic.csproj" />
   </ItemGroup>
 
   <ItemGroup>
