+++ GraphSynthSourceFiles/WpfCustomControlLibrary1/GraphSynth.GraphLayoutBasic.csproj
@@ -1,0 +1,13 @@
+﻿<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <UseWPF>true</UseWPF>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\GraphSynth.Base\GraphSynth.Base.csproj" />
+    <ProjectReference Include="..\GraphSynth.CustomControls\GraphSynth.CustomControls.csproj" />
+  </ItemGroup>
+
+</Project>
